+++ Antis-GameEngine(Arbeitsname)/MovementGenerator.h
@@ -29,7 +29,7 @@ public:
 
     D3DXVECTOR2 Move2DWithoutCollision(int x, int y, UINT uiMSECTime);
     /*D3DXVECTOR2 Move2D(int x, int y, UINT uiMSECTime);*/
-    void UpdateMovement(const UINT uiCurTime, const UINT uiDiff);
+    void UpdateMovement(const ULONGLONG uiCurTime, const UINT uiDiff);
     void ClearMovement();
     void RemoveMovementCommand(sMoveCommand* pCommand);
     inline bool IsMoveCommandListEmpty() { return m_lMoveCommands.empty(); }
